refactor: `@hooks/useFollowingsQuery`의 커스텀 훅을 `@hooks/user`로 이동 (#296)
