+++ client/src/components/FollowerUserContextMenu/index.tsx
@@ -3,8 +3,10 @@ import type { FC } from 'react';
 
 import defaultErrorHandler from '@errors/defaultErrorHandler';
 import { UserPlusIcon } from '@heroicons/react/20/solid';
-import { useInvalidateFollowingsQuery } from '@hooks/useFollowingsQuery';
-import { useFollowingMutation } from '@hooks/user';
+import {
+  useInvalidateFollowingsQuery,
+  useFollowingMutation,
+} from '@hooks/user';
 import { useRootStore } from '@stores/rootStore';
 import React from 'react';
