+++ client/src/components/FollowerUserSearchResult/index.tsx
@@ -5,7 +5,7 @@ import FollowerUserContextMenu from '@components/FollowerUserContextMenu';
 import UserItem from '@components/UserItem';
 import UserList from '@components/UserList';
 import { EllipsisHorizontalIcon } from '@heroicons/react/20/solid';
-import { useFollowingsMapQuery } from '@hooks/useFollowingsQuery';
+import { useFollowingsMapQuery } from '@hooks/user';
 import { useRootStore } from '@stores/rootStore';
 import React from 'react';
 import Scrollbars from 'react-custom-scrollbars-2';
